+++ bin/check-ec2-filter.rb
@@ -107,11 +107,11 @@ class EC2Filter < Sensu::Plugin::Check::CLI
 
     filter = Filter.parse(config[:filter])
 
-    if filter.empty?
-      options = {}
-    else
-      options = { filters: filter }
-    end
+    options = if filter.empty?
+                {}
+              else
+                { filters: filter }
+              end
 
     data = client.describe_instances(options)