+++ bin/check-elb-health-sdk.rb
@@ -92,11 +92,11 @@ class ELBHealth < Sensu::Plugin::Check::CLI
 
   def check_health(elb)
     unhealthy_instances = {}
-    if config[:instances]
-      instance_health_hash = elb.instances.health(config[:instances])
-    else
-      instance_health_hash = elb.instances.health
-    end
+    instance_health_hash = if config[:instances]
+                             elb.instances.health(config[:instances])
+                           else
+                             elb.instances.health
+                           end
     instance_health_hash.each do |instance_health|
       if instance_health[:state] != 'InService'
         unhealthy_instances[instance_health[:instance].id] = instance_health[:state]