+++ bin/check-instance-events.rb
@@ -72,10 +72,8 @@ class CheckInstanceEvents < Sensu::Plugin::Check::CLI
     aws_config = {}
 
     if config[:use_iam_role].nil?
-      aws_config.merge!(
-        access_key_id: config[:aws_access_key],
-        secret_access_key: config[:aws_secret_access_key]
-      )
+      aws_config[:access_key_id] = config[:aws_access_key]
+      aws_config[:secret_access_key] = config[:aws_secret_access_key]
     end
 
     ec2 = AWS::EC2::Client.new(aws_config.merge!(region: config[:aws_region]))