+++ bin/metrics-autoscaling-instance-count.rb
@@ -68,11 +68,11 @@ class AutoScalingInstanceCountMetrics < Sensu::Plugin::Metric::CLI::Graphite
   end
 
   def run
-    if config[:scheme] == ''
-      graphitepath = "#{config[:groupname]}.autoscaling.instance_count"
-    else
-      graphitepath = config[:scheme]
-    end
+    graphitepath = if config[:scheme] == ''
+                     "#{config[:groupname]}.autoscaling.instance_count"
+                   else
+                     config[:scheme]
+                   end
     begin
       as = AWS::AutoScaling.new aws_config
       count = as.groups[config[:groupname]].auto_scaling_instances.map(&:lifecycle_state).count('InService')