+++ bin/metrics-sqs.rb
@@ -68,11 +68,11 @@ class SQSMetrics < Sensu::Plugin::Metric::CLI::Graphite
   end
 
   def run
-    if config[:scheme] == ''
-      scheme = "aws.sqs.queue.#{config[:queue].tr('-', '_')}.message_count"
-    else
-      scheme = config[:scheme]
-    end
+    scheme = if config[:scheme] == ''
+               "aws.sqs.queue.#{config[:queue].tr('-', '_')}.message_count"
+             else
+               config[:scheme]
+             end
 
     begin
       sqs = AWS::SQS.new aws_config